Backend/requirements.txt
# FastAPI and Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0

# HTTP Clients and Requests
httpx==0.25.2
requests==2.31.0

# Database
pymongo==4.6.0

# Authentication and Security
pyjwt==2.8.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# Data Validation and Models
pydantic[email]==2.5.0

# Environment Variables
python-dotenv==1.0.0

# Date and Time Utilities
python-dateutil==2.8.2

# Async Support
asyncio-mqtt==0.13.0  # Optional for async operations

# Development Dependencies (Optional)
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
flake8==6.1.0

# Additional Utilities
typing-extensions==4.8.0